Single Player mode is now working again fully.  I had to do some magic for the GameState initialization due to the order of events in the Unity engine.  Also, the CameraController script is disabled until activated by the initialization.  I have made some changes to the coloring and materials of the walls and vehicles to try and get the vehicles to look better and the blue walls to not blend in as much with the background.  It's going to need some aditional work.

+++ Assets/Scripts/Global/GameSettings.cs
@@ -11,8 +11,8 @@ namespace Global
 {
     class GameSettings : NetworkBehaviour
     {
-        public float numAI = 0;
-        public float colorChoice = 0;
+        public int numAI = 0;
+        public int colorChoice = 0;
         
         // Use this for initialization
         void Start()
@@ -32,8 +32,8 @@ namespace Global
 
         public void setSettings()
         {
-            numAI = GameObject.FindGameObjectWithTag(GlobalTags.NUM_AI_SLIDER).GetComponent<Slider>().value;
-            colorChoice = GameObject.FindGameObjectWithTag(GlobalTags.COLOR_SLIDER).GetComponent<Slider>().value;
+            numAI = (int) GameObject.FindGameObjectWithTag(GlobalTags.NUM_AI_SLIDER).GetComponent<Slider>().value;
+            colorChoice = (int) GameObject.FindGameObjectWithTag(GlobalTags.COLOR_SLIDER).GetComponent<Slider>().value;
         }
     }
 }
